removing lot's of old issue reporter stuffs
