+++ src/vs/workbench/contrib/issue/common/issue.ts
@@ -6,7 +6,6 @@
 import { UriComponents } from '../../../../base/common/uri.js';
 import { ISandboxConfiguration } from '../../../../base/parts/sandbox/common/sandboxTypes.js';
 import { createDecorator } from '../../../../platform/instantiation/common/instantiation.js';
-import { OldIssueReporterData } from '../../../../platform/issue/common/issue.js';
 
 // Since data sent through the service is serialized to JSON, functions will be lost, so Color objects
 // should not be sent as their 'toString' method will be stripped. Instead convert to strings before sending.
@@ -107,16 +106,6 @@ export interface ProcessExplorerData extends WindowData {
 	styles: ProcessExplorerStyles;
 	platform: string;
 	applicationName: string;
-}
-
-export interface IssueReporterWindowConfiguration extends ISandboxConfiguration {
-	disableExtensions: boolean;
-	data: IssueReporterData | OldIssueReporterData;
-	os: {
-		type: string;
-		arch: string;
-		release: string;
-	};
 }
 
 export interface ProcessExplorerWindowConfiguration extends ISandboxConfiguration {
