--- src/vs/workbench/contrib/issue/electron-sandbox/issueReporterService2.ts
@@ -1,297 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-import { $, reset } from '../../../../base/browser/dom.js';
-import { CancellationError } from '../../../../base/common/errors.js';
-import { IProductConfiguration } from '../../../../base/common/product.js';
-import { URI } from '../../../../base/common/uri.js';
-import { localize } from '../../../../nls.js';
-import { isRemoteDiagnosticError } from '../../../../platform/diagnostics/common/diagnostics.js';
-import { IProcessMainService } from '../../../../platform/issue/common/issue.js';
-import { INativeHostService } from '../../../../platform/native/common/native.js';
-import { IThemeService } from '../../../../platform/theme/common/themeService.js';
-import { applyZoom } from '../../../../platform/window/electron-sandbox/window.js';
-import { BaseIssueReporterService } from '../browser/baseIssueReporterService.js';
-import { IssueReporterData as IssueReporterModelData } from '../browser/issueReporterModel.js';
-import { IIssueFormService, IssueReporterData, IssueType } from '../common/issue.js';
-
-// GitHub has let us know that we could up our limit here to 8k. We chose 7500 to play it safe.
-// ref https://github.com/microsoft/vscode/issues/159191
-const MAX_URL_LENGTH = 7500;
-
-
-export class IssueReporter2 extends BaseIssueReporterService {
-	private readonly processMainService: IProcessMainService;
-	constructor(
-		disableExtensions: boolean,
-		data: IssueReporterData,
-		os: {
-			type: string;
-			arch: string;
-			release: string;
-		},
-		product: IProductConfiguration,
-		window: Window,
-		@INativeHostService private readonly nativeHostService: INativeHostService,
-		@IIssueFormService issueFormService: IIssueFormService,
-		@IProcessMainService processMainService: IProcessMainService,
-		@IThemeService themeService: IThemeService
-	) {
-		super(disableExtensions, data, os, product, window, false, issueFormService, themeService);
-		this.processMainService = processMainService;
-		this.processMainService.$getSystemInfo().then(info => {
-			this.issueReporterModel.update({ systemInfo: info });
-			this.receivedSystemInfo = true;
-
-			this.updateSystemInfo(this.issueReporterModel.getData());
-			this.updatePreviewButtonState();
-		});
-		if (this.data.issueType === IssueType.PerformanceIssue) {
-			this.processMainService.$getPerformanceInfo().then(info => {
-				this.updatePerformanceInfo(info as Partial<IssueReporterData>);
-			});
-		}
-
-		this.setEventHandlers();
-		applyZoom(this.data.zoomLevel, this.window);
-		this.updateExperimentsInfo(this.data.experiments);
-		this.updateRestrictedMode(this.data.restrictedMode);
-		this.updateUnsupportedMode(this.data.isUnsupported);
-	}
-
-	public override setEventHandlers(): void {
-		super.setEventHandlers();
-
-		this.addEventListener('issue-type', 'change', (event: Event) => {
-			const issueType = parseInt((<HTMLInputElement>event.target).value);
-			this.issueReporterModel.update({ issueType: issueType });
-			if (issueType === IssueType.PerformanceIssue && !this.receivedPerformanceInfo) {
-				this.processMainService.$getPerformanceInfo().then(info => {
-					this.updatePerformanceInfo(info as Partial<IssueReporterData>);
-				});
-			}
-
-			// Resets placeholder
-			const descriptionTextArea = <HTMLInputElement>this.getElementById('issue-title');
-			if (descriptionTextArea) {
-				descriptionTextArea.placeholder = localize('undefinedPlaceholder', "Please enter a title");
-			}
-
-			this.updatePreviewButtonState();
-			this.setSourceOptions();
-			this.render();
-		});
-	}
-
-	public override async submitToGitHub(issueTitle: string, issueBody: string, gitHubDetails: { owner: string; repositoryName: string }): Promise<boolean> {
-		const url = `https://api.github.com/repos/${gitHubDetails.owner}/${gitHubDetails.repositoryName}/issues`;
-		const init = {
-			method: 'POST',
-			body: JSON.stringify({
-				title: issueTitle,
-				body: issueBody
-			}),
-			headers: new Headers({
-				'Content-Type': 'application/json',
-				'Authorization': `Bearer ${this.data.githubAccessToken}`
-			})
-		};
-
-		const response = await fetch(url, init);
-		if (!response.ok) {
-			console.error('Invalid GitHub URL provided.');
-			return false;
-		}
-		const result = await response.json();
-		await this.nativeHostService.openExternal(result.html_url);
-		this.close();
-		return true;
-	}
-
-	public override async createIssue(): Promise<boolean> {
-		const selectedExtension = this.issueReporterModel.getData().selectedExtension;
-		const hasUri = this.nonGitHubIssueUrl;
-		// Short circuit if the extension provides a custom issue handler
-		if (hasUri) {
-			const url = this.getExtensionBugsUrl();
-			if (url) {
-				this.hasBeenSubmitted = true;
-				await this.nativeHostService.openExternal(url);
-				return true;
-			}
-		}
-
-		if (!this.validateInputs()) {
-			// If inputs are invalid, set focus to the first one and add listeners on them
-			// to detect further changes
-			const invalidInput = this.window.document.getElementsByClassName('invalid-input');
-			if (invalidInput.length) {
-				(<HTMLInputElement>invalidInput[0]).focus();
-			}
-
-			this.addEventListener('issue-title', 'input', _ => {
-				this.validateInput('issue-title');
-			});
-
-			this.addEventListener('description', 'input', _ => {
-				this.validateInput('description');
-			});
-
-			this.addEventListener('issue-source', 'change', _ => {
-				this.validateInput('issue-source');
-			});
-
-			if (this.issueReporterModel.fileOnExtension()) {
-				this.addEventListener('extension-selector', 'change', _ => {
-					this.validateInput('extension-selector');
-					this.validateInput('description');
-				});
-			}
-
-			return false;
-		}
-
-		this.hasBeenSubmitted = true;
-
-		const issueTitle = (<HTMLInputElement>this.getElementById('issue-title')).value;
-		const issueBody = this.issueReporterModel.serialize();
-
-		let issueUrl = this.getIssueUrl();
-		if (!issueUrl) {
-			console.error('No issue url found');
-			return false;
-		}
-
-		if (selectedExtension?.uri) {
-			const uri = URI.revive(selectedExtension.uri);
-			issueUrl = uri.toString();
-		}
-
-		const gitHubDetails = this.parseGitHubUrl(issueUrl);
-
-		const baseUrl = this.getIssueUrlWithTitle((<HTMLInputElement>this.getElementById('issue-title')).value, issueUrl);
-		let url = baseUrl + `&body=${encodeURIComponent(issueBody)}`;
-
-		if (url.length > MAX_URL_LENGTH) {
-			try {
-				url = await this.writeToClipboard(baseUrl, issueBody);
-			} catch (_) {
-				console.error('Writing to clipboard failed');
-				return false;
-			}
-		} else if (this.data.githubAccessToken && gitHubDetails) {
-			return this.submitToGitHub(issueTitle, issueBody, gitHubDetails);
-		}
-
-		await this.nativeHostService.openExternal(url);
-		return true;
-	}
-
-	public override async writeToClipboard(baseUrl: string, issueBody: string): Promise<string> {
-		const shouldWrite = await this.issueFormService.showClipboardDialog();
-		if (!shouldWrite) {
-			throw new CancellationError();
-		}
-
-		await this.nativeHostService.writeClipboardText(issueBody);
-
-		return baseUrl + `&body=${encodeURIComponent(localize('pasteData', "We have written the needed data into your clipboard because it was too large to send. Please paste."))}`;
-	}
-
-	private updateSystemInfo(state: IssueReporterModelData) {
-		const target = this.window.document.querySelector<HTMLElement>('.block-system .block-info');
-
-		if (target) {
-			const systemInfo = state.systemInfo!;
-			const renderedDataTable = $('table', undefined,
-				$('tr', undefined,
-					$('td', undefined, 'CPUs'),
-					$('td', undefined, systemInfo.cpus || '')
-				),
-				$('tr', undefined,
-					$('td', undefined, 'GPU Status' as string),
-					$('td', undefined, Object.keys(systemInfo.gpuStatus).map(key => `${key}: ${systemInfo.gpuStatus[key]}`).join('\n'))
-				),
-				$('tr', undefined,
-					$('td', undefined, 'Load (avg)' as string),
-					$('td', undefined, systemInfo.load || '')
-				),
-				$('tr', undefined,
-					$('td', undefined, 'Memory (System)' as string),
-					$('td', undefined, systemInfo.memory)
-				),
-				$('tr', undefined,
-					$('td', undefined, 'Process Argv' as string),
-					$('td', undefined, systemInfo.processArgs)
-				),
-				$('tr', undefined,
-					$('td', undefined, 'Screen Reader' as string),
-					$('td', undefined, systemInfo.screenReader)
-				),
-				$('tr', undefined,
-					$('td', undefined, 'VM'),
-					$('td', undefined, systemInfo.vmHint)
-				)
-			);
-			reset(target, renderedDataTable);
-
-			systemInfo.remoteData.forEach(remote => {
-				target.appendChild($<HTMLHRElement>('hr'));
-				if (isRemoteDiagnosticError(remote)) {
-					const remoteDataTable = $('table', undefined,
-						$('tr', undefined,
-							$('td', undefined, 'Remote'),
-							$('td', undefined, remote.hostName)
-						),
-						$('tr', undefined,
-							$('td', undefined, ''),
-							$('td', undefined, remote.errorMessage)
-						)
-					);
-					target.appendChild(remoteDataTable);
-				} else {
-					const remoteDataTable = $('table', undefined,
-						$('tr', undefined,
-							$('td', undefined, 'Remote'),
-							$('td', undefined, remote.latency ? `${remote.hostName} (latency: ${remote.latency.current.toFixed(2)}ms last, ${remote.latency.average.toFixed(2)}ms average)` : remote.hostName)
-						),
-						$('tr', undefined,
-							$('td', undefined, 'OS'),
-							$('td', undefined, remote.machineInfo.os)
-						),
-						$('tr', undefined,
-							$('td', undefined, 'CPUs'),
-							$('td', undefined, remote.machineInfo.cpus || '')
-						),
-						$('tr', undefined,
-							$('td', undefined, 'Memory (System)' as string),
-							$('td', undefined, remote.machineInfo.memory)
-						),
-						$('tr', undefined,
-							$('td', undefined, 'VM'),
-							$('td', undefined, remote.machineInfo.vmHint)
-						)
-					);
-					target.appendChild(remoteDataTable);
-				}
-			});
-		}
-	}
-
-	private updateRestrictedMode(restrictedMode: boolean) {
-		this.issueReporterModel.update({ restrictedMode });
-	}
-
-	private updateUnsupportedMode(isUnsupported: boolean) {
-		this.issueReporterModel.update({ isUnsupported });
-	}
-
-	private updateExperimentsInfo(experimentInfo: string | undefined) {
-		this.issueReporterModel.update({ experimentInfo });
-		const target = this.window.document.querySelector<HTMLElement>('.block-experiments .block-info');
-		if (target) {
-			target.textContent = experimentInfo ? experimentInfo : localize('noCurrentExperiments', "No current experiments.");
-		}
-	}
-}
