+++ src/vs/workbench/contrib/issue/electron-sandbox/processService.ts
@@ -6,7 +6,7 @@
 import { getZoomLevel } from '../../../../base/browser/browser.js';
 import { platform } from '../../../../base/common/process.js';
 import { InstantiationType, registerSingleton } from '../../../../platform/instantiation/common/extensions.js';
-import { IProcessMainService, ProcessExplorerData } from '../../../../platform/issue/common/issue.js';
+import { IProcessMainService, ProcessExplorerData } from '../../../../platform/process/common/process.js';
 import { IProductService } from '../../../../platform/product/common/productService.js';
 import { activeContrastBorder, editorBackground, editorForeground, listActiveSelectionBackground, listActiveSelectionForeground, listFocusBackground, listFocusForeground, listFocusOutline, listHoverBackground, listHoverForeground, scrollbarShadow, scrollbarSliderActiveBackground, scrollbarSliderBackground, scrollbarSliderHoverBackground } from '../../../../platform/theme/common/colorRegistry.js';
 import { IColorTheme, IThemeService } from '../../../../platform/theme/common/themeService.js';
